requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-image>=0.18.0
pillow>=8.0.0
jupyter>=1.0.0
